WTF was I thinking with those tests?
I took some test conditions for the GSVD a little too far. I was trying to "help" the user but instead I just ended up being a stupid dummy.

Thus I should relax the condition of LW & RW. If an empty or identical matrix is input... that's now the user's problem (in part because it's actually necessary/important for, e.g., MCA).

Though with identical diagonal elements there is a way around to just multiply by a constant. That can be dealt with later.