Some updates & changes
The benchmarks were quite important here. The SVD wants more rows than columns. When it has that, it goes faster. But there is a slow own to then swap variables (e.g., $u an $v) back. So: in practice it is good to transpose but not make that decision for the user.