a wild amount of minor things and a new lesson learned:
devtools::check() vs. R CMD check, both with --as-cran do not give the same results (perhaps this is a problem I have where they point to different R versions, but I suspect not).

also that git problem is back where RStudio/Windows/git go into a death spiral of  a lock file and setting a folder as partial read only.